08_django/10_DRF/00_json_response/requirements.txt
asgiref==3.3.4
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
Faker==8.1.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
